requirements.txt
PyYAML==6.0
gspread==5.3.2
pandas==1.4.2
retry==0.9.2